libm: comparison of integers of different signs
../common/m9x/roundl.c:148:16: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
                        if (xx.i[0] < i)
                            ~~~~~~~ ^ ~
1 error generated.

../common/m9x/nexttowardl.c:95:5: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                                DEC(px)
                                ^~~~~~~
../common/m9x/nexttowardl.c:70:17: note: expanded from macro 'DEC'
                        if (--px[n2] == 0xffffffff) \
                            ~~~~~~~~ ^  ~~~~~~~~~~
../common/m9x/nexttowardl.c:100:5: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                                DEC(px)
                                ^~~~~~~
../common/m9x/nexttowardl.c:70:17: note: expanded from macro 'DEC'
                        if (--px[n2] == 0xffffffff) \
                            ~~~~~~~~ ^  ~~~~~~~~~~
2 errors generated.

../common/LD/nextafterl.c:95:5: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                                DEC(px)
                                ^~~~~~~
../common/LD/nextafterl.c:70:17: note: expanded from macro 'DEC'
                        if (--px[n2] == 0xffffffff) \
                            ~~~~~~~~ ^  ~~~~~~~~~~
../common/LD/nextafterl.c:100:5: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                                DEC(px)
                                ^~~~~~~
../common/LD/nextafterl.c:70:17: note: expanded from macro 'DEC'
                        if (--px[n2] == 0xffffffff) \
                            ~~~~~~~~ ^  ~~~~~~~~~~
2 errors generated.

../common/R/nextafterf.c:48:28: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        if (ix == iy || (ix | iy) == 0x80000000)
                         ~~~~~~~  ^  ~~~~~~~~~~
1 error generated.

../common/C/pow.c:208:25: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                                if ((j << (52 - k)) == ly)
                                     ~~~~~~~~~~~~~  ^  ~~
../common/C/pow.c:314:11: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                if (!(j == 0xc090cc00 && pz[LOWORD] == 0))      /* z < -1075 */
                      ~ ^  ~~~~~~~~~~
2 errors generated.

../common/C/exp2.c:70:11: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                        if (ix == 0xfff00000 && ((int *)&x)[LOWORD] == 0)
                            ~~ ^  ~~~~~~~~~~
1 error generated.

../common/C/exp10.c:78:11: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                        if (ix == 0xfff00000 && ((int *)&x)[LOWORD] == 0)
                            ~~ ^  ~~~~~~~~~~
1 error generated.

../common/C/exp.c:327:11: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                        if (ix == 0xfff00000 && ((int *)&x)[LOWORD] == 0)
                            ~~ ^  ~~~~~~~~~~
1 error generated.

../common/m9x/tgamma.c:1585:24: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                        if ((j << (52 - k)) == lx)
                             ~~~~~~~~~~~~~  ^  ~~
1 error generated.

../common/m9x/llroundl.c:149:16: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
                        if (xx.i[0] < i)
                            ~~~~~~~ ^ ~
1 error generated.

../common/m9x/__fex_sym.c:140:37: error: comparison of integers of different signs: 'off_t' (aka 'long') and 'Elf32_Off' (aka 'unsigned long') [-Werror,-Wsign-compare]
        if (lseek(fd, h.e_phoff, SEEK_SET) != h.e_phoff ||
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~
../common/m9x/__fex_sym.c:152:37: error: comparison of integers of different signs: 'off_t' (aka 'long') and 'Elf32_Off' (aka 'unsigned long') [-Werror,-Wsign-compare]
        if (lseek(fd, h.e_shoff, SEEK_SET) != h.e_shoff ||
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~
../common/m9x/__fex_sym.c:195:62: error: comparison of integers of different signs: 'ssize_t' (aka 'int') and 'Elf32_Word' (aka 'unsigned long') [-Werror,-Wsign-compare]
                read(fd, stbuf + sh[i].sh_size, sh[sh[i].sh_link].sh_size) !=
                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^
../common/m9x/__fex_sym.c:193:52: error: comparison of integers of different signs: 'off_t' (aka 'long') and 'Elf32_Off' (aka 'unsigned long') [-Werror,-Wsign-compare]
                lseek(fd, sh[sh[i].sh_link].sh_offset, SEEK_SET) !=
                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^
../common/m9x/__fex_sym.c:192:34: error: comparison of integers of different signs: 'ssize_t' (aka 'int') and 'Elf32_Word' (aka 'unsigned long') [-Werror,-Wsign-compare]
                read(fd, stbuf, sh[i].sh_size) != sh[i].sh_size ||
                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~
../common/m9x/__fex_sym.c:191:43: error: comparison of integers of different signs: 'off_t' (aka 'long') and 'Elf32_Off' (aka 'unsigned long') [-Werror,-Wsign-compare]
        if (lseek(fd, sh[i].sh_offset, SEEK_SET) != sh[i].sh_offset ||
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~
6 errors generated.

../common/m9x/__fex_i386.c:224:24: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        else if (*(1+(int*)x) == 0x80000000 && *(int*)x == 0)
                 ~~~~~~~~~~~~ ^  ~~~~~~~~~~

../common/LD/hypotl.c:116:14: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                        if (px[1] == 0x80000000 && px[0] == 0)
                            ~~~~~ ^  ~~~~~~~~~~
../common/LD/hypotl.c:118:30: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                        else if (ny == k && py[1] == 0x80000000 && py[0] == 0)
                                            ~~~~~ ^  ~~~~~~~~~~
2 errors generated.

../common/R/expf.c:362:11: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                        if (hx == 0xff800000)
                            ~~ ^  ~~~~~~~~~~
1 error generated.

+++ usr/src/lib/libm/common/C/exp10.c
@@ -75,7 +75,7 @@ exp10(double x) {
 
 	if (hx >= 0x4074a000) {	/* |x| >= 330 or x is nan */
 		if (hx >= 0x7ff00000) {	/* x is inf or nan */
-			if (ix == 0xfff00000 && ((int *)&x)[LOWORD] == 0)
+			if (ix == (int)0xfff00000 && ((int *)&x)[LOWORD] == 0)
 				return (zero);
 			return (x * x);
 		}
