+++ usr/src/lib/libm/common/m9x/llroundl.c
@@ -146,7 +146,7 @@ llroundl(long double x) {
 		} else {
 			i = 1 << (0x403d - ex);
 			xx.i[0] += i;
-			if (xx.i[0] < i)
+			if ((int)xx.i[0] < i)
 				xx.i[1]++;
 			xx.i[0] &= ~(i | (i - 1));
 		}
